A second clippy sweep, with three lints held back
`--fix` across the workspace: redundant closures (65, the bulk of it),
needless borrows, `map_or` that wants `is_some_and`, `unwrap_or_default`,
`clone` on Copy, needless returns. 709 warnings to 602, deduplicated by
(lint, file, line). Suite green, 48 test binaries.

Three lints were excluded with `-A`, for reasons rather than taste:

- `unused_imports` — last sweep's `--fix` deleted four imports used only
  under `#[cfg(test)]`, and `cargo build` stayed green while `cargo test`
  would not compile. Those are back with comments; automatic removal would
  take them out again.
- `manual_clamp` — not equivalent for floats. `x.max(lo).min(hi)` on NaN
  returns `lo`, because `f32::max` prefers the non-NaN operand;
  `x.clamp(lo, hi)` returns NaN. All nine sites are f32 geometry — scroll
  offsets, viewport sizes, cell metrics — where the current form quietly
  sanitises NaN and the "fix" would let it propagate into layout. Worth
  reading individually, not rewriting in bulk.
- `ptr_arg` — changes signatures, so it edits callers too.

One import did go: `use libc;`, flagged by `single_component_path_imports`
and redundant since Rust 2018.

On measuring this: a warning count taken straight after an incremental
build is worthless. The first count here came back 97, because cargo skips
unchanged crates and clippy emits nothing for them — a 6x undercount that
reads as spectacular progress. The real figure needs every file touched
first so all crates re-check. Same trap as `cmd | tail` reporting tail's
exit status: the number is real, it just answers a different question than
the one being asked.

+++ crates/cli/src/fetch/sftp.rs
@@ -770,7 +770,7 @@ fn discover_files(
         // Since filter
         if let Some(since) = since_ts {
             match mtime {
-                Some(t) if (t as u64) < since => continue,
+                Some(t) if t < since => continue,
                 None => continue,
                 _ => {}
             }
@@ -779,7 +779,7 @@ fn discover_files(
         // Min-age filter
         match mtime {
             Some(t) => {
-                let age = now.saturating_sub(t as u64);
+                let age = now.saturating_sub(t);
                 if age < min_age {
                     continue;
                 }
@@ -814,7 +814,7 @@ fn discover_files(
             path: full_path,
             filename,
             size,
-            mtime: mtime.map(|t| t as u64),
+            mtime: mtime.map(|t| t),
         });
     }
